requirements.txt
crewai>=0.22.8
litellm>=1.35.14
requests
PyYAML
